requirements.txt
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
pypdf>=4.0.0
arxiv==2.1.0
loguru==0.7.2
tenacity==8.3.0
discord.py==2.4.0
google-genai==0.8.0
openai==1.40.0

# RAG local embedding & vector store
numpy>=1.26.0
transformers>=4.51.0
safetensors>=0.4.3
faiss-cpu>=1.8.0.post1
torch>=2.2.0

# HTML parsing for ar5iv
beautifulsoup4>=4.12.3
lxml>=5.2.2
